requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
python-dotenv==1.0.1
tabulate==0.9.0
PyYAML==6.0.1
requests==2.31.0 